:hammer: Resolve cyclic dependencies
