+++ services/blockchain-connector/shared/sdk/endpoints.js
@@ -25,9 +25,8 @@ const {
 	getNodeInfo,
 	getSystemMetadata,
 } = require('./endpoints_1');
-const { timeoutMessage, getApiClient, invokeEndpoint, invokeEndpointProxy } = require('./client');
+const { timeoutMessage, getApiClient, invokeEndpoint } = require('./client');
 const { getGenesisHeight, getGenesisBlockID, getGenesisBlock } = require('./genesisBlock');
-const { formatEvent } = require('./formatter');
 
 const getConnectedPeers = async () => {
 	try {
@@ -241,7 +240,6 @@ const dryRunTransaction = async ({ transaction, skipVerify }) => {
 	try {
 		const apiClient = await getApiClient();
 		const response = await apiClient._channel.invoke('txpool_dryRunTransaction', { transaction, skipVerify });
-		response.events = response.events.map(event => formatEvent(event));
 		return response;
 	} catch (err) {
 		if (err.message.includes(timeoutMessage)) {
@@ -265,7 +263,6 @@ const getGenerators = async () => {
 
 module.exports = {
 	invokeEndpoint,
-	invokeEndpointProxy,
 	getSchemas,
 	getRegisteredActions,
 	getRegisteredEvents,
